April 21, 2014
Added defcoin wallet artwork and modified the scripts to create
defcoin wallets. The text color of the private keys was lightened
to make up for the dark colors of the wallet art.

The gpg signature remains unmodified at this point, in the hopes
that these changes will be pulled into the upstream branch.

	- allan